Add loop analysis, RPO merger, and global value numbering
Major backend refactor:
- Introduce LoopTree for loop detection and nesting in CFG
- Add RPO merger for linearizing IL after SSA lowering
- Replace CSE with global value numbering (GVN) pass
- Enhance CFG with dominance intervals and DFS traversal
- Update optimizer pipeline to use GVN, loop analysis, and RPO merge
- Improve Number class, jump table, and AST utilities
- Fix bugs in inlining and algebraic simplifier
- Remove obsolete common_sub_expression files
- Update example library and clean up submodules

+++ utils/include/utils/bump_allocator.hpp
@@ -17,7 +17,7 @@ namespace HXSL
 			uint32_t blockSize;
 			uint32_t used;
 
-			Block(Block* prev, size_t blockSize) : prev(prev), blockSize(blockSize), used(0)
+			Block(Block* prev, uint32_t blockSize) : prev(prev), blockSize(blockSize), used(0)
 			{	
 			}
